+++ src/main/java/com/quodex/dineboard/repository/PlanRepository.java
@@ -3,5 +3,5 @@ package com.quodex.dineboard.repository;
 import com.quodex.dineboard.model.Plan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlanRepository extends JpaRepository<Plan, Long> {
+public interface PlanRepository extends JpaRepository<Plan, Integer> {
 }
